% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapviz_methods.R
\name{dimnames.shapviz}
\alias{dimnames.shapviz}
\title{Dimnames of "shapviz" Object}
\usage{
\method{dimnames}{shapviz}(x)
}
\arguments{
\item{x}{An object of class "shapviz".}
}
\value{
Dimnames of the SHAP matrix.
}
\description{
Dimnames of "shapviz" Object
}
\examples{
S <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X <- data.frame(x = c("a", "b"), y = c(100, 10))
x <- shapviz(S, X, baseline = 4)
dimnames(x)

# Implies colnames()
colnames(x)
}
\seealso{
\code{\link{shapviz}}.
}
