% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_importance.R
\name{format_aligned}
\alias{format_aligned}
\title{Aligned Number Formatter}
\usage{
format_aligned(x, digits = 4, scientific = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector to be formatted.}

\item{digits}{Number of digits to be shown, including the zero in front of the decimal dot.}

\item{scientific}{Should scientific formatting be applied? Default is \code{FALSE}.}

\item{...}{Further arguments passed to \code{format()}.}
}
\value{
A character vector of formatted numbers.
}
\description{
Formats a vector of positive numbers in a way that the resulting strings are
all of the same length and the decimal points (if there are any) are aligned.
}
\examples{
x <- c(10, 1, 0.1)
format_aligned(x)
}
