% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapviz_methods.R
\name{is.shapviz}
\alias{is.shapviz}
\title{Check for shapviz}
\usage{
is.shapviz(object)
}
\arguments{
\item{object}{An R object.}
}
\value{
Returns \code{TRUE} if \code{object} has "\code{shapviz}" among its classes, and \code{FALSE} otherwise.
}
\description{
Is object of class "shapviz"?
}
\examples{
S <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X <- data.frame(x = c("a", "b"), y = c(100, 10))
shp <- shapviz(S, X)
is.shapviz(shp)
is.shapviz("a")
}
