% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapviz_methods.R
\name{+.shapviz}
\alias{+.shapviz}
\title{Concatenates "shapviz" Object}
\usage{
\method{+}{shapviz}(x, ...)
}
\arguments{
\item{x}{An object of class "shapviz".}

\item{...}{"shapviz" objects to be concatenated}
}
\value{
A new object of class "shapviz".
}
\description{
Use standard plus operator to concatenate a number of "shapviz" objects.
}
\examples{
S1 <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
S2 <- matrix(c(-1, 1, 1, -1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X1 <- data.frame(x = c("a", "b"), y = c(100, 10))
X2 <- data.frame(x = c("b", "a"), y = c(100, 10))
x1 <- shapviz(S1, X1, baseline = 4)
x2 <- shapviz(S2, X2, baseline = 4)
x1 + x2
}
\seealso{
\code{\link{shapviz}}.
}
