% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mshapviz}
\alias{print.mshapviz}
\title{Prints "mshapviz" Object}
\usage{
\method{print}{mshapviz}(x, ...)
}
\arguments{
\item{x}{An object of class "mshapviz".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Prints "mshapviz" Object
}
\examples{
TODO
}
\seealso{
\code{\link{mshapviz}}.
}
