% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.shapviz}
\alias{print.shapviz}
\title{Prints "shapviz" Object}
\usage{
\method{print}{shapviz}(x, ...)
}
\arguments{
\item{x}{An object of class "shapviz".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Prints "shapviz" Object
}
\examples{
S <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X <- data.frame(x = c("a", "b"), y = c(100, 10))
x <- shapviz(S, X, baseline = 4)
x
}
\seealso{
\code{\link[=shapviz]{shapviz()}}
}
