% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[.mshapviz}
\alias{[.mshapviz}
\title{Subsets "mshapviz" Object}
\usage{
\method{[}{mshapviz}(x, i, j, ...)
}
\arguments{
\item{x}{An object of class "mshapviz".}

\item{i}{Row subsetting.}

\item{j}{Column subsetting.}

\item{...}{Currently unused.}
}
\value{
A new object of class "mshapviz".
}
\description{
Use standard square bracket subsetting to select rows and/or columns of
SHAP values, feature values, and SHAP interaction values of a "mshapviz" object.
}
\examples{
S1 <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
S2 <- matrix(c(-1, 1, 1, -1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X1 <- data.frame(x = c("a", "b"), y = c(100, 10))
X2 <- data.frame(x = c("b", "a"), y = c(100, 10))
x1 <- shapviz(S1, X1, baseline = 4)
x2 <- shapviz(S2, X2, baseline = 4)
x <- c(Model_1 = x1, Model_2 = x2)
x
x[1, ]
}
\seealso{
\code{\link{shapviz}}.
}
