% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.shapviz}
\alias{summary.shapviz}
\title{Summarizes "shapviz" Object}
\usage{
\method{summary}{shapviz}(object, n = 2L, ...)
}
\arguments{
\item{object}{An object of class "shapviz".}

\item{n}{Maximum number of rows of SHAP values and feature values to show.}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Summarizes "shapviz" Object
}
\examples{
S <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X <- data.frame(x = c("a", "b"), y = c(100, 10))
object <- shapviz(S, X, baseline = 4)
summary(object)
}
\seealso{
\code{\link{shapviz}}
}
