% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_importance.R
\name{format_max}
\alias{format_max}
\title{Number Formatter}
\usage{
format_max(x, digits = 4, scientific = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector to be formatted.}

\item{digits}{Number of digits before and after the decimal separator to be shown.}

\item{scientific}{Should scientific formatting be applied? Default is \code{FALSE}.}

\item{...}{Further arguments passed to \code{prettyNum()}, e.g., \code{big.mark = "'"}.}
}
\value{
A character vector of formatted numbers.
}
\description{
Formats a vector of positive numbers in a way that the largest value determines the
number of digits.
}
\examples{
x <- c(10, 1, 0.1)
format_max(x)
}
