% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.mshapviz}
\alias{is.mshapviz}
\title{Check for mshapviz}
\usage{
is.mshapviz(object)
}
\arguments{
\item{object}{An R object.}
}
\value{
Returns \code{TRUE} if \code{object} has "\code{mshapviz}" among its classes, and \code{FALSE} otherwise.
}
\description{
Is object of class "mshapviz"?
}
\examples{
TODO
}
