% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{+.shapviz}
\alias{+.shapviz}
\alias{+.mshapviz}
\title{Rowbinds two "shapviz" Objects}
\usage{
\method{+}{shapviz}(e1, e2)

\method{+}{mshapviz}(e1, e2)
}
\arguments{
\item{e1}{The first object of class "shapviz".}

\item{e2}{The second object of class "shapviz".}
}
\value{
A new object of class "shapviz".
}
\description{
Rowbinds two "shapviz" objects using \code{+}.
}
\examples{
S <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X <- data.frame(x = c("a", "b"), y = c(100, 10))
s1 <- shapviz(S, X, baseline = 4)[1]
s2 <- shapviz(S, X, baseline = 4)[2]
s <- s1 + s2
s
# mshapviz
S <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X <- data.frame(x = c("a", "b"), y = c(100, 10))
s1 <- shapviz(S, X, baseline = 4)[1L]
s2 <- shapviz(S, X, baseline = 4)[2L]
s <- mshapviz(c(shp1 = s1, shp2 = s2))
s + s

}
\seealso{
\code{\link[=shapviz]{shapviz()}}, \code{\link[=rbind.shapviz]{rbind.shapviz()}}
}
