% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{+.shapviz}
\alias{+.shapviz}
\title{Rowbinds two "shapviz" Objects}
\usage{
\method{+}{shapviz}(e1, e2)
}
\arguments{
\item{e1}{The first object of class "shapviz".}

\item{e2}{The second object of class "shapviz".}
}
\value{
A new object of class "shapviz".
}
\description{
Rowbinds two "shapviz" objects using \code{+}.
}
\examples{
S1 <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
S2 <- matrix(c(-1, 1, 1, -1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X1 <- data.frame(x = c("a", "b"), y = c(100, 10))
X2 <- data.frame(x = c("b", "a"), y = c(100, 10))
x1 <- shapviz(S1, X1, baseline = 4)
x2 <- shapviz(S2, X2, baseline = 4)
x1 + x2
}
\seealso{
\code{\link{shapviz}}.
}
