% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mshapviz}
\alias{print.mshapviz}
\title{Prints "mshapviz" Object}
\usage{
\method{print}{mshapviz}(x, ...)
}
\arguments{
\item{x}{An object of class "mshapviz".}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Prints "mshapviz" Object
}
\examples{
S <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X <- data.frame(x = c("a", "b"), y = c(100, 10))
s1 <- shapviz(S, X, baseline = 4)
s2 <- shapviz(S, X, baseline = 4)
x <- c(s1 = s1, s2 = s2)
x
}
\seealso{
\code{\link{shapviz}}.
}
