% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapviz_methods.R
\name{rbind.shapviz}
\alias{rbind.shapviz}
\title{Concatenates "shapviz" Object}
\usage{
\method{rbind}{shapviz}(...)
}
\arguments{
\item{...}{"shapviz" objects to be concatenated}
}
\value{
A new object of class "shapviz".
}
\description{
It is based on the \code{+.shapviz} operator
}
\examples{
S1 <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
S2 <- matrix(c(-1, 1, 1, -1), ncol = 2, dimnames = list(NULL, c("x", "y")))
X1 <- data.frame(x = c("a", "b"), y = c(100, 10))
X2 <- data.frame(x = c("b", "a"), y = c(100, 10))
x1 <- shapviz(S1, X1, baseline = 4)
x2 <- shapviz(S2, X2, baseline = 4)
rbind(x1, x2)
}
\seealso{
\code{\link{shapviz}}.
}
