% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapviz-package.R
\docType{package}
\name{shapviz-package}
\alias{shapviz-package}
\alias{_PACKAGE}
\title{shapviz: SHAP Visualizations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Visualizations for SHAP (SHapley Additive exPlanations), such as waterfall plots, force plots, various types of importance plots, dependence plots, and interaction plots. These plots act on a 'shapviz' object created from a matrix of SHAP values and a corresponding feature dataset. Wrappers for the R packages 'xgboost', 'lightgbm', 'fastshap', 'shapr', 'h2o', 'treeshap', 'DALEX', 'kernelshap', and 'permshap' are added for convenience. By separating visualization and computation, it is possible to display factor variables in graphs, even if the SHAP values are calculated by a model that requires numerical features. The plots are inspired by those provided by the 'shap' package in Python, but there is no dependency on it.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ModelOriented/shapviz}
  \item Report bugs at \url{https://github.com/ModelOriented/shapviz/issues}
}

}
\author{
\strong{Maintainer}: Michael Mayer \email{mayermichael79@gmail.com}

Other contributors:
\itemize{
  \item Adrian Stando \email{adrian.j.stando@gmail.com} [contributor]
}

}
